feat: rename old `/background` page to `/timeline` page
